update to synchro with main repo

+++ PostProcess/submit_job_automated_new_multiple_vcfs.sh
@@ -27,14 +27,17 @@ current_working_directory=$(realpath ${15})
 gene_proximity=$(realpath ${16})
 
 email=${17}
+echo -e "MAIL: $email" 
 echo -e "CPU used: $ncpus" 
 
 
 log=$output_folder/log.txt
 touch $log
+#echo -e 'Job\tStart\t'$(date) > $log
+start_time='Job\tStart\t'$(date)
 
-output=$output_folder/output.txt
-touch $output
+# output=$output_folder/output.txt
+# touch $output
 
 rm -f $output_folder/queue.txt
 #for vcf_f in "${vcf_list[@]}"; 
@@ -62,9 +65,9 @@ do
 	pam_name=$(basename $4)
 	annotation_name=$(basename $5)
 
-
-	echo -e 'Job\tStart\t'$(date) > $log
-	echo -e 'Job\tStart\t'$(date) >&2 
+	echo -e $start_time > $log
+	# echo -e 'Job\tStart\t'$(date) > $log
+	# echo -e 'Job\tStart\t'$(date) >&2 
 
 	unset real_chroms
 	declare -a real_chroms
@@ -134,7 +137,7 @@ do
 		cd "$current_working_directory/Genomes"
 		if ! [ -d "$current_working_directory/Genomes/${ref_name}+${vcf_name}" ]; then
 			echo -e 'Add-variants\tStart\t'$(date) >> $log
-			echo -e 'Add-variants\tStart\t'$(date) >&2
+			# echo -e 'Add-variants\tStart\t'$(date) >&2
 			echo -e "Adding variants"
 			crispritz.py add-variants "$vcf_folder/" "$ref_folder/" "true"
 			#if ! [ -d "${ref_name}+${vcf_name}" ]; then
@@ -154,12 +157,12 @@ do
 				mkdir "genome_library/${true_pam}_2_${ref_name}+${vcf_name}_INDELS"
 			fi
 			echo -e 'Add-variants\tEnd\t'$(date) >> $log
-			echo -e 'Add-variants\tEnd\t'$(date) >&2
+			# echo -e 'Add-variants\tEnd\t'$(date) >&2
 			echo -e 'Indexing Indels\tStart\t'$(date) >> $log
-			echo -e 'Indexing Indels\tStart\t'$(date) >&2
+			# echo -e 'Indexing Indels\tStart\t'$(date) >&2
 			${starting_dir}/./pool_index_indels.py "$current_working_directory/Genomes/variants_genome/" "$pam_file" $true_pam $ref_name $vcf_name $ncpus
 			echo -e 'Indexing Indels\tEnd\t'$(date) >> $log
-			echo -e 'Indexing Indels\tEnd\t'$(date) >&2
+			# echo -e 'Indexing Indels\tEnd\t'$(date) >&2
 			if ! [ -d $current_working_directory/Genomes/${ref_name}+${vcf_name}_INDELS ]; then
 				mkdir $current_working_directory/Genomes/${ref_name}+${vcf_name}_INDELS
 			fi
@@ -177,10 +180,10 @@ do
 	if [ "$vcf_name" != "_" ]; then
 		if ! [ -d "$current_working_directory/genome_library/${true_pam}_2_${ref_name}+${vcf_name}_INDELS" ]; then
 			echo -e 'Indexing Indels\tStart\t'$(date) >> $log
-			echo -e 'Indexing Indels\tStart\t'$(date) >&2
+			# echo -e 'Indexing Indels\tStart\t'$(date) >&2
 			${starting_dir}/./pool_index_indels.py "$current_working_directory/Genomes/${ref_name}+${vcf_name}_INDELS/" "$pam_file" $true_pam $ref_name $vcf_name $ncpus
 			echo -e 'Indexing Indels\tEnd\t'$(date) >> $log
-			echo -e 'Indexing Indels\tEnd\t'$(date) >&2
+			# echo -e 'Indexing Indels\tEnd\t'$(date) >&2
 		fi
 	fi
 
@@ -194,13 +197,13 @@ do
 			if ! [ $bMax -gt 1 ]; then
 				if ! [ -d "$current_working_directory/genome_library/${true_pam}_1_${ref_name}" ]; then
 					echo -e 'Index-genome Reference\tStart\t'$(date) >> $log	
-					echo -e 'Index-genome Reference\tStart\t'$(date) >&2	
-					echo -e 'Indexing_Reference' > $output
+					# echo -e 'Index-genome Reference\tStart\t'$(date) >&2	
+					# echo -e 'Indexing_Reference' > $output
 					echo -e "Indexing reference genome"
 					crispritz.py index-genome "$ref_name" "$ref_folder/" "$pam_file" -bMax $bMax -th $ncpus
 					pid_index_ref=$!
 					echo -e 'Index-genome Reference\tEnd\t'$(date) >> $log	
-					echo -e 'Index-genome Reference\tEnd\t'$(date) >&2	
+					# echo -e 'Index-genome Reference\tEnd\t'$(date) >&2	
 					idx_ref="$current_working_directory/genome_library/${true_pam}_${bMax}_${ref_name}"
 				else
 					echo -e "Reference Index already present"
@@ -208,25 +211,25 @@ do
 				fi
 			else
 				echo -e 'Index-genome Reference\tStart\t'$(date) >> $log	
-				echo -e 'Index-genome Reference\tStart\t'$(date) >&2	
-				echo -e 'Indexing_Reference' > $output
+				# echo -e 'Index-genome Reference\tStart\t'$(date) >&2	
+				# echo -e 'Indexing_Reference' > $output
 				echo -e "Indexing reference genome"
 				crispritz.py index-genome "$ref_name" "$ref_folder/" "$pam_file" -bMax $bMax -th $ncpus
 				pid_index_ref=$!
 				echo -e 'Index-genome Reference\tEnd\t'$(date) >> $log	
-				echo -e 'Index-genome Reference\tEnd\t'$(date) >&2	
+				# echo -e 'Index-genome Reference\tEnd\t'$(date) >&2	
 				idx_ref="$current_working_directory/genome_library/${true_pam}_${bMax}_${ref_name}"
 			fi
 		else
 			echo -e "Reference Index already present"
 			echo -e 'Index-genome Reference\tEnd\t'$(date) >> $log	
-			echo -e 'Index-genome Reference\tEnd\t'$(date) >&2	
+			# echo -e 'Index-genome Reference\tEnd\t'$(date) >&2	
 			idx_ref="$current_working_directory/genome_library/${true_pam}_2_${ref_name}"
 		fi
 	else
 		echo -e "Reference Index already present"
 		echo -e 'Index-genome Reference\tEnd\t'$(date) >> $log	
-		echo -e 'Index-genome Reference\tEnd\t'$(date) >&2
+		# echo -e 'Index-genome Reference\tEnd\t'$(date) >&2
 		idx_ref="$current_working_directory/genome_library/${true_pam}_${bMax}_${ref_name}"
 	fi
 
@@ -237,13 +240,13 @@ do
 				if ! [ $bMax -gt 1 ]; then
 					if ! [ -d "$current_working_directory/genome_library/${true_pam}_1_${ref_name}+${vcf_name}" ]; then
 						echo -e 'Index-genome Variant\tStart\t'$(date) >> $log	
-						echo -e 'Index-genome Variant\tStart\t'$(date) >&2	
-						echo -e 'Indexing_Enriched' > $output
+						# echo -e 'Index-genome Variant\tStart\t'$(date) >&2	
+						# echo -e 'Indexing_Enriched' > $output
 						echo -e "Indexing variant genome"
 						crispritz.py index-genome "${ref_name}+${vcf_name}" "$current_working_directory/Genomes/${ref_name}+${vcf_name}/" "$pam_file" -bMax $bMax -th $ncpus #${ref_folder%/}+${vcf_name}/
 						pid_index_var=$!
 						echo -e 'Index-genome Variant\tEnd\t'$(date) >> $log	
-						echo -e 'Index-genome Variant\tEnd\t'$(date) >&2	
+						# echo -e 'Index-genome Variant\tEnd\t'$(date) >&2	
 						idx_var="$current_working_directory/genome_library/${true_pam}_${bMax}_${ref_name}+${vcf_name}"
 					else
 						echo -e "Variant Index already present"
@@ -251,25 +254,25 @@ do
 					fi
 				else
 					echo -e 'Index-genome Variant\tStart\t'$(date) >> $log	
-					echo -e 'Index-genome Variant\tStart\t'$(date) >&2	
-					echo -e 'Indexing_Enriched' > $output
+					# echo -e 'Index-genome Variant\tStart\t'$(date) >&2	
+					# echo -e 'Indexing_Enriched' > $output
 					echo -e "Indexing variant genome"
 					crispritz.py index-genome "${ref_name}+${vcf_name}" "$current_working_directory/Genomes/${ref_name}+${vcf_name}/" "$pam_file" -bMax $bMax -th $ncpus
 					pid_index_ref=$!
 					echo -e 'Index-genome Variant\tEnd\t'$(date) >> $log	
-					echo -e 'Index-genome Variant\tEnd\t'$(date) >&2	
+					# echo -e 'Index-genome Variant\tEnd\t'$(date) >&2	
 					idx_var="$current_working_directory/genome_library/${true_pam}_${bMax}_${ref_name}+${vcf_name}"
 				fi
 			else
 				echo -e "Variant Index already present"
 				echo -e 'Index-genome Variant\tEnd\t'$(date) >> $log
-				echo -e 'Index-genome Variant\tEnd\t'$(date) >&2
+				# echo -e 'Index-genome Variant\tEnd\t'$(date) >&2
 				idx_var="$current_working_directory/genome_library/${true_pam}_2_${ref_name}+${vcf_name}"
 			fi
 		else
 			echo -e "Variant Index already present"
 			echo -e 'Index-genome Variant\tEnd\t'$(date) >> $log
-			echo -e 'Index-genome Variant\tEnd\t'$(date) >&2
+			# echo -e 'Index-genome Variant\tEnd\t'$(date) >&2
 			idx_var="$current_working_directory/genome_library/${true_pam}_${bMax}_${ref_name}+${vcf_name}"
 		fi	
 	fi
@@ -278,8 +281,8 @@ do
 	echo $idx_ref
 	if ! [ -f "$output_folder/crispritz_targets/${ref_name}_${pam_name}_${guide_name}_${mm}_${bDNA}_${bRNA}.targets.txt" ]; then
 		echo -e 'Search Reference\tStart\t'$(date) >> $log	
-		echo -e 'Search Reference\tStart\t'$(date) >&2
-		echo -e 'Search Reference' >  $output
+		# echo -e 'Search Reference\tStart\t'$(date) >&2
+		# echo -e 'Search Reference' >  $output
 		crispritz.py search $idx_ref "$pam_file" "$guide_file" "${ref_name}_${pam_name}_${guide_name}_${mm}_${bDNA}_${bRNA}" -index -mm $mm -bDNA $bDNA -bRNA $bRNA -t  -th $(expr $ncpus / 4) &
 		pid_search_ref=$!
 	else
@@ -289,12 +292,12 @@ do
 	if [ "$vcf_name" != "_" ]; then
 		if ! [ -f "$output_folder/crispritz_targets/${ref_name}+${vcf_name}_${pam_name}_${guide_name}_${mm}_${bDNA}_${bRNA}.targets.txt" ]; then
 			echo -e 'Search Variant\tStart\t'$(date) >> $log	
-			echo -e 'Search Variant\tStart\t'$(date) >&2	
-			echo -e 'Search Variant' >  $output
+			# echo -e 'Search Variant\tStart\t'$(date) >&2	
+			# echo -e 'Search Variant' >  $output
 			crispritz.py search "$idx_var" "$pam_file" "$guide_file" "${ref_name}+${vcf_name}_${pam_name}_${guide_name}_${mm}_${bDNA}_${bRNA}" -index -mm $mm -bDNA $bDNA -bRNA $bRNA -t  -th $(expr $ncpus / 4) -var
 			mv "${ref_name}+${vcf_name}_${pam_name}_${guide_name}_${mm}_${bDNA}_${bRNA}.targets.txt" "$output_folder/crispritz_targets"
 			echo -e 'Search Variant\tEnd\t'$(date) >> $log	
-			echo -e 'Search Variant\tEnd\t'$(date) >&2	
+			# echo -e 'Search Variant\tEnd\t'$(date) >&2	
 		else
 			echo -e "Search for variant already done"
 		fi
@@ -302,13 +305,13 @@ do
 		if ! [ -f "$output_folder/crispritz_targets/indels_${ref_name}+${vcf_name}_${pam_name}_${guide_name}_${mm}_${bDNA}_${bRNA}.targets.txt" ]; then
 			echo -e "Search INDELs Start"
 			echo -e 'Search INDELs\tStart\t'$(date) >> $log	
-			echo -e 'Search INDELs\tStart\t'$(date) >&2
+			# echo -e 'Search INDELs\tStart\t'$(date) >&2
 			cd $starting_dir
 			./pool_search_indels.py "$ref_folder" "$vcf_folder" "$vcf_name" "$guide_file" "$pam_file" $bMax $mm $bDNA $bRNA "$output_folder" $true_pam "$current_working_directory/"
 			mv "$output_folder/indels_${ref_name}+${vcf_name}_${pam_name}_${guide_name}_${mm}_${bDNA}_${bRNA}.targets.txt" "$output_folder/crispritz_targets"
 			echo -e "Search INDELs End"
 			echo -e 'Search INDELs\tEnd\t'$(date) >> $log
-			echo -e 'Search INDELs\tEnd\t'$(date) >&2
+			# echo -e 'Search INDELs\tEnd\t'$(date) >&2
 		else
 			echo -e "Search INDELs already done"
 		fi
@@ -319,7 +322,7 @@ do
 		sleep 100
 	done
 	echo -e 'Search Reference\tEnd\t'$(date) >> $log	
-	echo -e 'Search Reference\tEnd\t'$(date) >&2	
+	# echo -e 'Search Reference\tEnd\t'$(date) >&2	
 
 	if [ -f "$output_folder/${ref_name}_${pam_name}_${guide_name}_${mm}_${bDNA}_${bRNA}.targets.txt" ]; then
 		mv "$output_folder/${ref_name}_${pam_name}_${guide_name}_${mm}_${bDNA}_${bRNA}.targets.txt" "$output_folder/crispritz_targets"
@@ -334,10 +337,10 @@ do
 
 	echo -e "Start post-analysis"
 
-	echo -e 'Post analysis' >  $output
+	# echo -e 'Post analysis' >  $output
 	if [ "$vcf_name" != "_" ]; then
 		echo -e 'Post-analysis SNPs\tStart\t'$(date) >> $log	
-		echo -e 'Post-analysis SNPs\tStart\t'$(date) >&2	
+		# echo -e 'Post-analysis SNPs\tStart\t'$(date) >&2	
 		final_res="$output_folder/final_results_$(basename ${output_folder}).bestMerge.txt"
 		final_res_alt="$output_folder/final_results_$(basename ${output_folder}).altMerge.txt"
 		if ! [ -f "$final_res" ]; then
@@ -349,8 +352,8 @@ do
 		
 		./pool_post_analisi_snp.py $output_folder $ref_folder $vcf_name $guide_file $mm $bDNA $bRNA $annotation_file $pam_file $dict_folder $final_res $final_res_alt $ncpus
 		
-		echo -e 'Post-analysis SNPs\tEnd\t'$(date) >> $log	
-		echo -e 'Post-analysis SNPs\tEnd\t'$(date) >&2	
+			
+		# echo -e 'Post-analysis SNPs\tEnd\t'$(date) >&2	
 		for key in "${real_chroms[@]}"
 		do
 			echo "Concatenating $key"
@@ -359,9 +362,12 @@ do
 			rm "$output_folder/${ref_name}+${vcf_name}_${pam_name}_${guide_name}_${annotation_name}_${mm}_${bDNA}_${bRNA}_$key.bestMerge.txt"
 			# rm "$output_folder/${ref_name}+${vcf_name}_${pam_name}_${guide_name}_${annotation_name}_${mm}_${bDNA}_${bRNA}_$key.altMerge.txt"
 		done
+
+		echo -e 'Post-analysis SNPs\tEnd\t'$(date) >> $log
+
 	else
 		echo -e 'Post-analysis\tStart\t'$(date) >> $log	
-		echo -e 'Post-analysis\tStart\t'$(date) >&2	
+		# echo -e 'Post-analysis\tStart\t'$(date) >&2	
 		final_res="$output_folder/final_results_$(basename ${output_folder}).bestMerge.txt"
 		final_res_alt="$output_folder/final_results_$(basename ${output_folder}).altMerge.txt"
 		if ! [ -f "$final_res" ]; then
@@ -381,7 +387,7 @@ do
 			# rm "$output_folder/${ref_name}+${vcf_name}_${pam_name}_${guide_name}_${annotation_name}_${mm}_${bDNA}_${bRNA}_$key.altMerge.txt"
 		done
 		echo -e 'Post-analysis\tEnd\t'$(date) >> $log	
-		echo -e 'Post-analysis\tEnd\t'$(date) >&2	
+		# echo -e 'Post-analysis\tEnd\t'$(date) >&2	
 
 	fi
 
@@ -390,10 +396,10 @@ do
 		cd "$starting_dir"
 		
 		echo -e 'Post-analysis INDELs\tStart\t'$(date) >> $log	
-		echo -e 'Post-analysis INDELs\tStart\t'$(date) >&2	
+		# echo -e 'Post-analysis INDELs\tStart\t'$(date) >&2	
 		./pool_post_analisi_indel.py $output_folder $ref_folder $vcf_folder $guide_file $mm $bDNA $bRNA $annotation_file $pam_file "$current_working_directory/Dictionaries/" $final_res $final_res_alt $ncpus
-		echo -e 'Post-analysis INDELs\tEnd\t'$(date) >> $log	
-		echo -e 'Post-analysis INDELs\tEnd\t'$(date) >&2	
+			
+		# echo -e 'Post-analysis INDELs\tEnd\t'$(date) >&2	
 		for key in "${array_fake_chroms[@]}"
 		do	
 			echo "Concatenating $key"
@@ -402,6 +408,8 @@ do
 			rm "$output_folder/${key}_${pam_name}_${guide_name}_${annotation_name}_${mm}_${bDNA}_${bRNA}.bestMerge.txt"
 			rm "$output_folder/${key}_${pam_name}_${guide_name}_${annotation_name}_${mm}_${bDNA}_${bRNA}.altMerge.txt"
 		done
+
+		echo -e 'Post-analysis INDELs\tEnd\t'$(date) >> $log
 
 	fi
 done < $vcf_list
@@ -416,35 +424,42 @@ do
 	# tail -n +2 $samples >> "$output_folder/sampleID.txt"
 	grep -v '#' "${current_working_directory}/samplesIDs/$samples" >> "$output_folder/sampleID.txt"
 done < $sampleID
-sed -i 1i"#SAMPLE_ID\tPOPULATION_ID\tSUPERPOPULATION_ID\tGENDER" "$output_folder/sampleID.txt"
+if [ -f "$output_folder/sampleID.txt" ]; then
+	sed -i 1i"#SAMPLE_ID\tPOPULATION_ID\tSUPERPOPULATION_ID\tGENDER" "$output_folder/sampleID.txt"
+fi
 
 sampleID=$output_folder/sampleID.txt
 
 
-echo -e 'Merging targets' >  $output
-echo -e 'Merging Close Targets\tStart\t'$(date) >> $log	
-echo -e 'Merging Close Targets\tStart\t'$(date) >&2	
+# echo -e 'Merging targets' >  $output
+echo -e 'Merging Targets\tStart\t'$(date) >> $log
+#echo -e 'Merging Close Targets\tStart\t'$(date) >> $log	
 ./merge_close_targets_cfd.sh $final_res $final_res.trimmed $merge_t
 mv $final_res.trimmed $final_res
 mv $final_res.trimmed.discarded_samples $final_res_alt
 
-echo -e 'Merging Close Targets\tEnd\t'$(date) >> $log	
-echo -e 'Merging Close Targets\tEnd\t'$(date) >&2	
+# echo -e 'Merging Close Targets\tEnd\t'$(date) >> $log	
+# echo -e 'Merging Close Targets\tEnd\t'$(date) >&2	
 
-echo -e 'Merging Alternative Chromosomes\tStart\t'$(date) >> $log	
-echo -e 'Merging Alternative Chromosomes\tStart\t'$(date) >&2	
+# echo -e 'Merging Alternative Chromosomes\tStart\t'$(date) >> $log	
+# echo -e 'Merging Alternative Chromosomes\tStart\t'$(date) >&2	
 ./merge_alt_chr.sh $final_res $final_res.chr_merged
 
-echo -e 'Merging Alternative Chromosomes\tEnd\t'$(date) >> $log	
-echo -e 'Merging Alternative Chromosomes\tEnd\t'$(date) >&2	
+# echo -e 'Merging Alternative Chromosomes\tEnd\t'$(date) >> $log	
+# echo -e 'Merging Alternative Chromosomes\tEnd\t'$(date) >&2	
+echo -e 'Merging Targets\tEnd\t'$(date) >> $log
 
 mv $final_res.chr_merged $final_res
 
 sed -i '1 s/^.*$/#Bulge_type\tcrRNA\tDNA\tReference\tChromosome\tPosition\tCluster_Position\tDirection\tMismatches\tBulge_Size\tTotal\tPAM_gen\tVar_uniq\tSamples\tAnnotation_Type\tReal_Guide\trsID\tAF\tSNP\t#Seq_in_cluster\tCFD\tCFD_ref\tMMBLG_#Bulge_type\tMMBLG_crRNA\tMMBLG_DNA\tMMBLG_Reference\tMMBLG_Chromosome\tMMBLG_Position\tMMBLG_Cluster_Position\tMMBLG_Direction\tMMBLG_Mismatches\tMMBLG_Bulge_Size\tMMBLG_Total\tMMBLG_PAM_gen\tMMBLG_Var_uniq\tMMBLG_Samples\tMMBLG_Annotation_Type\tMMBLG_Real_Guide\tMMBLG_rsID\tMMBLG_AF\tMMBLG_SNP\tMMBLG_#Seq_in_cluster\tMMBLG_CFD\tMMBLG_CFD_ref/' "$final_res"
 sed -i '1 s/^.*$/#Bulge_type\tcrRNA\tDNA\tReference\tChromosome\tPosition\tCluster_Position\tDirection\tMismatches\tBulge_Size\tTotal\tPAM_gen\tVar_uniq\tSamples\tAnnotation_Type\tReal_Guide\trsID\tAF\tSNP\t#Seq_in_cluster\tCFD\tCFD_ref\tMMBLG_#Bulge_type\tMMBLG_crRNA\tMMBLG_DNA\tMMBLG_Reference\tMMBLG_Chromosome\tMMBLG_Position\tMMBLG_Cluster_Position\tMMBLG_Direction\tMMBLG_Mismatches\tMMBLG_Bulge_Size\tMMBLG_Total\tMMBLG_PAM_gen\tMMBLG_Var_uniq\tMMBLG_Samples\tMMBLG_Annotation_Type\tMMBLG_Real_Guide\tMMBLG_rsID\tMMBLG_AF\tMMBLG_SNP\tMMBLG_#Seq_in_cluster\tMMBLG_CFD\tMMBLG_CFD_ref/' "$final_res_alt"
 
+echo -e 'Annotating results\tStart\t'$(date) >> $log	
+# echo -e 'Annotating results\tStart\t'$(date) >&2	
 ./annotate_final_results.py $final_res $annotation_file $final_res.annotated
 ./annotate_final_results.py $final_res_alt $annotation_file $final_res_alt.annotated
+echo -e 'Annotating results\tEnd\t'$(date) >> $log	
+# echo -e 'Annotating results\tEnd\t'$(date) >&2	
 
 mv $final_res.annotated $final_res
 mv $final_res_alt.annotated $final_res_alt
@@ -456,11 +471,13 @@ if ! [ -d "$output_folder/cfd_graphs" ]; then
 fi
 ./assemble_cfd_graphs.py $output_folder
 mv $output_folder/snps.CFDGraph.txt $output_folder/cfd_graphs
-mv $output_folder/indels.CFDGraph.txt $output_folder/cfd_graphs
+if [ -f "$output_folder/indels.CFDGraph.txt" ]; then
+	mv $output_folder/indels.CFDGraph.txt $output_folder/cfd_graphs
+fi
 
-echo -e 'Creating images' >  $output
+# echo -e 'Creating images' >  $output
 echo -e 'Creating images\tStart\t'$(date) >> $log	
-echo -e 'Creating images\tStart\t'$(date) >&2	
+# echo -e 'Creating images\tStart\t'$(date) >&2	
 echo -e "Adding risk score"
 ./add_risk_score.py $final_res $final_res.risk
 mv "$final_res.risk" "${output_folder}/$(basename ${output_folder}).bestMerge.txt"
@@ -509,7 +526,7 @@ else
 	rm dummy.txt
 fi
 echo -e 'Creating images\tEnd\t'$(date) >> $log	
-echo -e 'Creating images\tEnd\t'$(date) >&2	
+# echo -e 'Creating images\tEnd\t'$(date) >&2	
 
 # if [ "$vcf_name" != "_" ]; then
 # 	cp $sampleID $output_folder/sampleID.txt
@@ -517,17 +534,17 @@ echo -e 'Creating images\tEnd\t'$(date) >&2
 
 echo -e 'Building database'
 echo -e 'Creating database\tStart\t'$(date) >> $log
-echo -e 'Creating database\tStart\t'$(date) >&2
+# echo -e 'Creating database\tStart\t'$(date) >&2
 if [ -f "${output_folder}/$(basename ${output_folder}).db" ]; then
 	rm "${output_folder}/$(basename ${output_folder}).db"
 fi
 python $starting_dir/db_creation.py "${output_folder}/$(basename ${output_folder}).bestMerge.txt" "${output_folder}/$(basename ${output_folder})"
 echo -e 'Creating database\tEnd\t'$(date) >> $log
-echo -e 'Creating database\tEnd\t'$(date) >&2
+# echo -e 'Creating database\tEnd\t'$(date) >&2
 
 echo $gene_proximity
 echo -e 'Integrating results\tStart\t'$(date) >> $log
-echo -e 'Integrating results\tStart\t'$(date) >&2
+# echo -e 'Integrating results\tStart\t'$(date) >&2
 echo >> $guide_file
 # while read guide;
 # do
@@ -561,10 +578,10 @@ if [ $6 != "_" ]; then
 fi
 
 echo -e 'Integrating results\tEnd\t'$(date) >> $log
-echo -e 'Integrating results\tEnd\t'$(date) >&2
+# echo -e 'Integrating results\tEnd\t'$(date) >&2
 echo -e 'Job\tDone\t'$(date) >> $log
-echo -e 'Job\tDone\t'$(date) >&2
-echo -e 'Job End' >  $output
+# echo -e 'Job\tDone\t'$(date) >&2
+# echo -e 'Job End' >  $output
 echo -e "JOB END"
 
 if [ "$email" != "" ]; then
